requirements.txt
alembic==0.9.10
asn1crypto==0.24.0
cffi==1.11.5
Django==1.11.4
dominate==2.3.1
flask-blueprint==1.2.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
pycparser==2.18
PyMySQL==0.9.0
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.4
SQLAlchemy==1.2.9
urllib3==1.23
visitor==0.1.3
WTForms==2.2.1
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
cryptography==2.2.2
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Session==0.3.1
idna==2.7
Mako==1.0.7
redis==2.10.6
requests==2.19.1
six==1.11.0


